Move and cleanup dialogs.
