+++ src/app/app.module.ts
@@ -22,15 +22,15 @@ import { ArchivedCarsComponent } from './components/archived-cars/archived-cars.
 import { CarCreateComponent } from './components/car-create/car-create.component';
 import { CarDeleteComponent } from './components/car-delete/car-delete.component';
 import { CarDetailViewComponent } from './components/car-detail-view/car-detail-view.component';
-import { CarDialogComponent } from './components/car-dialog/car-dialog.component';
 import { CarListComponent } from './components/car-list/car-list.component';
 import { CarUpdateComponent } from './components/car-update/car-update.component';
-import { ConfirmDialogComponent } from './components/confirm-dialog/confirm-dialog.component';
-import { FuelDialogComponent } from './components/fuel-dialog/fuel-dialog.component';
 import { LoadingComponent } from './components/loading/loading.component';
 import { LoginComponent } from './components/login/login.component';
 import { MainContainerComponent } from './components/main-container/main-container.component';
 import { ShellComponent } from './components/shell/shell.component';
+import { CarDialogComponent } from './dialogs/car-dialog/car-dialog.component';
+import { ConfirmDialogComponent } from './dialogs/confirm-dialog/confirm-dialog.component';
+import { FuelDialogComponent } from './dialogs/fuel-dialog/fuel-dialog.component';
 import { CarState } from './state/car.state';
 import { UserState } from './state/user.state';
 import { CarListViewComponent } from './views/car-list-view/car-list-view.component';
