+++ src/app/components/car-delete/car-delete.component.html
@@ -1,3 +1,1 @@
-<div clrDropdownItem (click)="deleteCar(car.id)">
-  <clr-icon shape="trash" class="is-error"></clr-icon> {{ 'Car.DeleteCar' | translate }}
-</div>
+<div clrDropdownItem (click)="deleteCar(car.id)"><clr-icon shape="trash" class="is-error"></clr-icon> {{ 'Car.Delete' | translate }}</div>
