+++ src/app/components/car-update/car-update.component.html
@@ -1,1 +1,1 @@
-<div clrDropdownItem (click)="updateCar()"><clr-icon shape="pencil"></clr-icon> {{ 'Car.UpdateCar' | translate }}</div>
+<div clrDropdownItem (click)="updateCar()"><clr-icon shape="pencil"></clr-icon> {{ 'Car.Update' | translate }}</div>
