+++ src/app/dialogs/confirm-dialog/confirm-dialog.component.scss
