+++ src/app/dialogs/fuel-dialog/fuel-dialog.component.html
@@ -1,5 +1,5 @@
 <clr-modal [(clrModalOpen)]="isOpen" [clrModalClosable]="false">
-  <h3 class="modal-title">{{ isNew ? ('Fuel.NewFuel' | translate) : ('Fuel.UpdateFuel' | translate) }}</h3>
+  <h3 class="modal-title">{{ isNew ? ('Fuel.Create' | translate) : ('Fuel.Update' | translate) }}</h3>
   <div class="modal-body">
     <form clrForm [formGroup]="form" clrLayout="vertical" class="full-width">
       <div class="clr-row">
